Updated all files - desc
Updated all the files to include a short description - and last modified date such that is needed in the appendices of the report.

Also removed a few files that are not used or have never been used.
Also do we need both points.h or could the one under \Chess\Points.h be removed